pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spclaw"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "python-dotenv>=1.0.1",
  "slack-bolt>=1.18.1",
  "pypdf>=5.2.0",
  "python-docx>=1.1.2",
  "python-pptx>=1.0.2",
  "youtube-transcript-api>=0.6.2",
  "yt-dlp>=2025.1.26",
  "matplotlib>=3.8.0",
  "numpy>=1.26.0",
  "pandas>=2.1.0",
  "openai>=1.0.0",
  "lancedb>=0.23.0",
  "yfinance>=0.2.40",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
spclaw = "spclaw.cli:main"
claw = "spclaw.cli:main"
